Show nextInt() can use with data type long and not otherwise

+++ ICPC Assiut  Sheets Version 3/Data type - Conditions - sh1/src/C.java
@@ -7,7 +7,7 @@ public class C {
     public static void main(String[] args) {
 
     long x , y;
-    x = input.nextInt();
+    x = input.nextLong();
     y = input.nextInt();
 
     System.out.println(x + " + " + y + " = " + (x+y));
